Optimize FusedMoe kernel V2
